# swagger.yaml
swagger: "2.0"
info:
  version: "2.0.0"
  title: "Hotel Audit System API"
  description: "Please consider that resources with a lock require a valid secretKey query parameter on the /v1 path (\"v1_auth\") and a valid JWT access token (\"v2_auth_query\" or \"v2_auth_header\") on the /v2 path."
  contact:
    email: "martin.biermann@hrs.com"
host: "api.hotel-audit.hrs.com"
...
schemes:
- "https"
paths:
...
  /v1/audits:
    get:
      tags:
      - "v1"
      deprecated: true
      summary: ""
      description: ""
      operationId: "getAudits"
      produces:
      - "application/json"
      parameters:
      - name: "hkeys"
        in: "query"
        description: "Comma-separated list of hkeys"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v1_auth: []
  /v1/audits/report:
    get:
      tags:
      - "v1"
      deprecated: true
      summary: "Pagination on Clean & Safe audits"
      description: "Paginated listing of Clean & Safe audit reports"
      operationId: "getAuditReports"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      - name: "where"
        in: "query"
        description: "Filter on updated date field of audit reports in format YYYY-MM-DD"
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v1_auth: []
  /v1/audits/report/touchlessstay:
    get:
      tags:
      - "v1"
      deprecated: true
      summary: "Pagination on Touchless Stay audits"
      description: "Paginated listing of Touchless Stay audit reports"
      operationId: "getTouchlessStayReports"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      - name: "where"
        in: "query"
        description: "Filter on updated date field of audit reports in format YYYY-MM-DD"
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v1_auth: []
...
  /v2/audits:
    get:
      tags:
      - "v2"
      summary: ""
      description: ""
      operationId: "getAuditsV2"
      produces:
      - "application/json"
      parameters:
      - name: "hkeys"
        in: "query"
        description: "Comma-separated list of hkeys"
        required: true
        type: "string"
      - name: "exclude"
        in: "query"
        description: "Comma-separated list of audit categories to be excluded from response. Audit categories are 'touchless','cleansafe','green', 'geosure' and 'checkin'."
        type: "string"
      - name: "include"
        in: "query"
        description: "Comma-separated list of audit categories to be included from response. Audit categories are 'touchless','cleansafe','green', 'geosure' and 'checkin'."
        type: "string"
      - name: "backfill"
        in: "query"
        description: "***For internal and educational purposes only - should not be used in public production environments.***\nOptional flag to provide calculated Green Stay record based on similar hotels, if no actual record exists for the hotel."
        required: false
        type: "boolean"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
    options:
      tags:
      - "v2"
      operationId: "corsGetAuditsV2"
      responses:
        '204':
          description: "Successful operation"
  /v2/audits/report:
    get:
      tags:
      - "v2"
      summary: "Pagination on Clean & Safe audits"
      description: "Paginated listing of Clean & Safe audit reports"
      operationId: "getAuditReportsV2"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      - name: "where"
        in: "query"
        description: "Filter on updated date field of audit reports in format YYYY-MM-DD"
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
  /v2/audits/report/touchlessstay:
    get:
      tags:
      - "v2"
      summary: "Pagination on Touchless Stay audits"
      description: "Paginated listing of Touchless Stay audit reports"
      operationId: "getTouchlessStayReportsV2"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      - name: "where"
        in: "query"
        description: "Filter on updated date field of audit reports in format YYYY-MM-DD"
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
  /v2/audits/report/green:
    get:
      tags:
      - "v2"
      summary: "Pagination on Green Stay audits"
      description: "Paginated listing of Green Stay audit reports"
      operationId: "getGreenStayAuditReportsV2"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      - name: "where"
        in: "query"
        description: "Filter on updated date field of audit reports in format YYYY-MM-DD"
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
  /v2/audits/report/green_tracking:
    get:
      tags:
      - "v2"
      summary: "Pagination on Green Stay tracking records"
      description: "Paginated listing of Green Stay tracking records"
      operationId: "getGreenStayTrackingReportsV2"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      - name: "where"
        in: "query"
        description: "Filter on updated date field of audit reports in format YYYY-MM-DD"
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
  /v2/audits/report/green_exceptions:
    get:
      tags:
      - "v2"
      summary: "Pagination on Green Stay exception records"
      description: "Paginated listing of Green Stay exception records"
      operationId: "getGreenStayExceptionReportsV2"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      - name: "where"
        in: "query"
        description: "Filter on updated date field of audit reports in format YYYY-MM-DD"
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
  /v2/audits/green/reports/{id}:
    get:
      tags:
      - "v2"
      summary: "Green Stay Audit Report"
      description: "Green Stay Audit Report"
      operationId: "getGreenAuditReport"

      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Report ID"
        type: "string"
        required: true
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
  /v2/audits/report/geosure:
    get:
      tags:
      - "v2"
      summary: "Pagination on GeoSure audits"
      description: "Paginated listing of GeoSure audit reports"
      operationId: "getGeoSureAuditReportsV2"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      - name: "where"
        in: "query"
        description: "Filter on updated date field of audit reports in format YYYY-MM-DD"
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
...
securityDefinitions:
  v1_auth:
    type: "apiKey"
    name: "secretKey"
    in: "query"
  v2_auth_query:
    type: "apiKey"
    name: "token"
    in: "query"
  v2_auth_header:
    type: "apiKey"
    name: "x-access-token"
    in: "header"
...